moved /v2 to /echtzeit
